Fixed Buffer.prototype.indexOf() on 32bits platforms.

+++ src/njs_buffer.c
@@ -2198,7 +2198,7 @@ encoding:
             }
 
         } else {
-            to -= str.length - 1;
+            to -= (int64_t) str.length - 1;
 
             if (from > to) {
                 goto done;
